Added remove_excess_elements to transform_to_keras_input and uncommented start of fit_LSTM
The output of transform_to_keras_input will now a) always be numpy arrays, b) not need any further processing and so c) can be fed directly to model.fit()

The previous solution performed more data processing inside fit_LSTM. I felt that this code was murky and hacky, so made it more clear.

Possible loss of performance by transforming to tf.data.Dataset and back to NumPy, but it's what I know and the first time I've done a project this size.